requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
matplotlib== 3.9.3
seaborn==0.13.0
pandas==2.1.4
numpy ==1.26.3
